@@ -4,61 +4,60 @@ description: An application can help you collect game data more easily via
 
 publish_to: "none"
 
-version: 0.0.5+1
+version: 0.1.0+1
 
 environment:
-  sdk: ">=2.18.1 <3.0.0"
+  sdk: ">=2.19.3 <3.0.0"
 
 dependencies:
   another_flushbar: 1.12.29
-  badges: 2.0.3
+  badges: 3.0.2
   bot_toast: 4.0.3
-  collection: 1.16.0
-  connectivity_plus: 3.0.2
+  collection: 1.17.0
+  connectivity_plus: 3.0.3
   dartx: 1.1.0
   dartz: 0.10.1
-  data_table_2: 2.3.8
-  dio: 4.0.6
-  drift: 2.2.0
+  data_table_2: 2.4.1
+  dio: 5.0.2
+  drift: 2.5.0
   easy_localization: 3.0.1
   ffi: 2.0.1
-  file_picker: 5.2.2
-  fl_chart: 0.55.2
-  fluent_ui: 4.0.3+1
+  file_picker: 5.2.6
+  fl_chart: 0.61.0
+  fluent_ui: 4.4.1
   flutter:
     sdk: flutter
-  flutter_riverpod: 2.1.1
-  flutter_screenutil: 5.6.0
-  font_awesome_flutter: 10.2.1
+  flutter_riverpod: 2.3.2
+  flutter_screenutil: 5.7.0
+  font_awesome_flutter: 10.4.0
   freezed_annotation: 2.2.0
-  go_router: 5.1.10
+  go_router: 6.2.0
   intl: 0.17.0
-  json_annotation: 4.7.0
-  logger: 1.1.0
-  modal_bottom_sheet: 2.1.2
-  package_info_plus: 3.0.2
+  json_annotation: 4.8.0
+  logger: 1.3.0
+  modal_bottom_sheet: 3.0.0-pre
+  package_info_plus: 3.0.3
   path: 1.8.2
-  path_provider: 2.0.11
-  pretty_dio_logger: 1.2.0-beta-1
-  retrofit: 3.3.1
-  riverpod: 2.1.1
-  shared_preferences: 2.0.15
-  sqlite3_flutter_libs: 0.5.11+1
+  path_provider: 2.0.13
+  pretty_dio_logger: 1.3.1
+  retrofit: 4.0.1
+  shared_preferences: 2.0.18
+  sqlite3_flutter_libs: 0.5.13
   time: 2.1.3
-  url_launcher: 6.1.6
+  url_launcher: 6.1.10
   webview_windows: 0.2.1
-  win32: 3.1.1
-  window_manager: 0.2.7
+  win32: 3.1.3
+  window_manager: 0.3.1
 
 dev_dependencies:
-  build_runner: 2.3.2
-  drift_dev: 2.2.0+1
+  build_runner: 2.3.3
+  drift_dev: 2.5.2
   flutter_test:
     sdk: flutter
-  freezed: 2.2.1
-  json_serializable: 6.5.4
-  lint: 1.10.0
-  retrofit_generator: 4.2.0
+  freezed: 2.3.2
+  json_serializable: 6.6.1
+  lint: 2.0.1
+  retrofit_generator: 6.0.0+1
 
 flutter:
   uses-material-design: true
