@@ -1,42 +1,62 @@
 import 'package:dartz/dartz.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../../core/enums/gacha_rule_type.dart';
 import '../../domain/gacha/gacha_stats.dart';
 import '../../domain/user/user.dart';
 import '../../infrastructure/gacha/gacha_repository.dart';
 import '../user/user_provider.dart';
 import 'params/get_gacha_stats_params.dart';
 
-final gachaStatsProvider = StateNotifierProvider.autoDispose<GachaStatsNotifier,
-    AsyncValue<GachaStats>>(
-  (ref) => GachaStatsNotifier(
-    ref.watch(userProvider.notifier),
+final gachaStatsProvider = StateNotifierProvider.autoDispose
+    .family<GachaStatsNotifier, AsyncValue<GachaStats>, String?>(
+  (ref, String? pool) => GachaStatsNotifier(
+    pool,
+    ref.watch(userProvider),
     ref.watch(gachaRepositoryProvider),
   ),
-  dependencies: [gachaRepositoryProvider],
+  dependencies: [
+    userProvider,
+    gachaRepositoryProvider,
+  ],
 );
 
 class GachaStatsNotifier extends StateNotifier<AsyncValue<GachaStats>> {
   GachaStatsNotifier(
-    this._userProvider,
+    this._pool,
+    this._userOption,
     this._repository,
-  ) : super(const AsyncValue.loading());
+  ) : super(const AsyncValue.loading()) {
+    _get();
+  }
 
-  final StateController<Option<User>> _userProvider;
+  final String? _pool;
+  final Option<User> _userOption;
   final GachaRepository _repository;
 
-  late GetGachaStatsParams _cachedParams;
+  late GetGachaStatsParams _params;
 
-  Future<void> get(GetGachaStatsParams params) async {
-    _userProvider.state.fold(
+  Future<void> _get() async {
+    final pool = _pool;
+    if (pool == null) {
+      // 全部寻访
+      _params = GetGachaStatsParams.all();
+    } else if (pool == GachaRuleType.normal.label) {
+      // 标准寻访
+      _params = GetGachaStatsParams.normal();
+    } else {
+      // 指定寻访
+      _params = GetGachaStatsParams.specific(pool);
+    }
+    _userOption.fold(
       () {},
       (user) async {
         state = const AsyncValue.loading();
         final failureOrStats = await _repository.getStats(
           user.uid,
-          pool: params.pool,
-          includeRuleTypes: params.includeRuleTypes,
-          excludeRuleTypes: params.excludeRuleTypes,
+          pool: _params.pool,
+          includeRuleTypes: _params.includeRuleTypes,
+          excludeRuleTypes: _params.excludeRuleTypes,
         );
         state = await AsyncValue.guard(
           () async => failureOrStats.fold(
@@ -44,10 +64,7 @@ class GachaStatsNotifier extends StateNotifier<AsyncValue<GachaStats>> {
             (stats) => stats,
           ),
         );
-        _cachedParams = params;
       },
     );
   }
-
-  Future<void> refresh() => get(_cachedParams);
 }
