@@ -13,7 +13,7 @@ class Pagination with _$Pagination {
 
   const Pagination._();
 
-  int get totalPage => (total / pageSize).ceil();
+  int get totalPage => (total / Constants.pageSize).ceil();
 
   bool get isLastPage => current >= totalPage;
 }
