@@ -66,9 +66,9 @@ class DiamondHistoryTable extends StatelessWidget {
                 showCheckboxColumn: false,
                 showFirstLastButtons: true,
                 availableRowsPerPage: const [
-                  pageSize,
-                  pageSize * 2,
-                  pageSize * 5,
+                  Constants.pageSize,
+                  Constants.pageSize * 2,
+                  Constants.pageSize * 5,
                 ],
                 onRowsPerPageChanged: (_) {},
                 wrapInCard: false,
@@ -98,7 +98,7 @@ class _DataTableSource extends material.DataTableSource {
         _buildTextCell('${index + 1}'),
         _buildTextCell(change.type),
         _buildTextCell(change.description, color: change.textColor),
-        _buildTextCell(change.after.toString()),
+        _buildTextCell('${change.after}'),
         _buildTextCell(change.operation),
         _buildTextCell(change.ts.dateTime.yMMMdHmsString),
       ],
