@@ -6,36 +6,43 @@ import '../../application/user/user_fetch_provider.dart';
 import '../core/common/widgets/app_error_view.dart';
 import 'widgets/index.dart';
 
-class GachaStatsPage extends ConsumerWidget {
+class GachaStatsPage extends StatelessWidget {
   const GachaStatsPage({super.key});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return ref.watch(userFetchProvider).when(
-          data: (_) => Padding(
-            padding: EdgeInsets.symmetric(vertical: 20.h, horizontal: 24.w),
-            child: Column(
-              children: [
-                _header,
-                SizedBox(height: 48.h),
-                const Expanded(child: GachaStatsView()),
-              ],
-            ),
-          ),
-          error: (_, __) => const AppErrorView(),
-          loading: () => const SizedBox.expand(
-            child: Center(child: ProgressRing()),
-          ),
-        );
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.symmetric(vertical: 20.h, horizontal: 24.w),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          Text('寻访统计', style: TextStyle(fontSize: 24.sp)),
+          SizedBox(height: 20.h),
+          _header,
+          SizedBox(height: 20.h),
+          Expanded(child: _buildStatesView()),
+        ],
+      ),
+    );
   }
 
   Widget get _header => Row(
-        children: [
-          const UserInfoView(),
-          const Spacer(),
-          const GachaPoolSelector(),
-          SizedBox(width: 80.w),
-          const GachaStatsExtraPanel(),
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: const [
+          GachaPoolSelector(),
+          GachaStatsExtraPanel(),
         ],
       );
+
+  Widget _buildStatesView() {
+    return Consumer(
+      builder: (_, ref, __) => ref.watch(userFetchProvider).when(
+            data: (_) => const GachaStatsView(),
+            error: (_, __) => const AppErrorView(),
+            loading: () => const SizedBox.expand(
+              child: Center(child: ProgressRing()),
+            ),
+          ),
+    );
+  }
 }
