@@ -1,32 +1,22 @@
-import 'package:badges/badges.dart';
-import 'package:dartx/dartx.dart';
-import 'package:data_table_2/data_table_2.dart';
-import 'package:fl_chart/fl_chart.dart';
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart' as material;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 
-import '../../../application/gacha/gacha_pie_chart_provider.dart';
+import '../../../application/gacha/gacha_pool_filter_provider.dart';
 import '../../../application/gacha/gacha_pool_selector_provider.dart';
 import '../../../application/gacha/gacha_provider.dart';
 import '../../../application/gacha/gacha_stats_provider.dart';
-import '../../../application/gacha/params/get_gacha_stats_params.dart';
-import '../../../core/constants/constants.dart';
 import '../../../core/enums/rarity.dart';
 import '../../../domain/gacha/gacha_stats.dart';
-import '../../../infrastructure/core/extensions/date_time_formatter.dart';
-import '../../core/common/widgets/app_badge.dart';
-import '../../core/common/widgets/app_empty_view.dart';
 import '../../core/common/widgets/app_error_view.dart';
-import '../../core/routing/route_params.dart';
-import '../../core/routing/router.dart';
 
-final _selectedPool = Provider.autoDispose(
-  (ref) => ref.watch(gachaPoolSelectorProvider).selectedPool,
-);
+final _filteredPools = Provider.autoDispose((ref) {
+  final selectedPool = ref.watch(gachaPoolSelectorProvider).selectedPool;
+  if (selectedPool == null) {
+    return ref.watch(gachaPoolSelectorProvider).pools;
+  }
+  return [selectedPool];
+});
 
 class GachaStatsView extends ConsumerWidget {
   const GachaStatsView({super.key});
@@ -48,9 +38,17 @@ class GachaStatsView extends ConsumerWidget {
             );
           },
           failure: (_) => const AppErrorView(),
-          success: (_) => const _StatsView(),
+          success: (_) => _contentView,
         );
   }
+
+  Widget get _contentView => Consumer(
+        builder: (_, ref, __) => ref.watch(gachaStatsProvider(null)).when(
+              data: (_) => const _StatsView(),
+              error: (_, __) => const AppErrorView(),
+              loading: () => const SizedBox.shrink(),
+            ),
+      );
 
   Widget _buildProgressBar({
     double? value,
@@ -68,397 +66,220 @@ class GachaStatsView extends ConsumerWidget {
       );
 }
 
-class _StatsView extends ConsumerStatefulWidget {
+class _StatsView extends ConsumerWidget {
   const _StatsView();
 
   @override
-  ConsumerState<_StatsView> createState() => _StatsViewState();
-}
-
-class _StatsViewState extends ConsumerState<_StatsView> {
-  @override
-  void initState() {
-    super.initState();
-    Future.delayed(Duration.zero, _fetch);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    _listenState();
-
-    final pool = ref.watch(_selectedPool);
-    return ref.watch(gachaStatsProvider).when(
-          data: (stats) => _PieChart(pool, stats),
-          error: (_, __) => const AppErrorView(),
-          loading: () => const SizedBox.shrink(),
-        );
-  }
-
-  void _fetch({String? pool}) {
-    late final GetGachaStatsParams params;
-    if (pool == null) {
-      // 标准寻访
-      params = GetGachaStatsParams.normal();
-    } else {
-      // 指定寻访
-      params = GetGachaStatsParams.pool(pool);
-    }
-    ref.read(gachaStatsProvider.notifier).get(params);
-  }
-
-  void _listenState() {
-    ref.listen(
-      _selectedPool,
-      (_, pool) => _fetch(pool: pool),
-    );
-  }
-}
-
-class _PieChart extends StatelessWidget {
-  const _PieChart(this.pool, this.stats);
-
-  final String? pool;
-  final GachaStats stats;
-
-  String get title => pool ?? '标准寻访';
-
-  AutoDisposeChangeNotifierProvider<GachaPieChartNotifier>
-      get _gachaPieChartProvider => gachaPieChartProvider(title);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 400.w,
-      padding: EdgeInsets.symmetric(vertical: 24.h),
-      child: Column(
-        children: [
-          _titleLabel,
-          SizedBox(height: 28.h),
-          _indicators,
-          Container(
-            padding: EdgeInsets.all(4.w),
-            width: 320.w,
-            height: 320.h,
-            child: _chart,
-          ),
-          _pullDateRangeLabel,
-          SizedBox(height: 16.h),
-          _totalPullsLabel,
-          SizedBox(height: 4.h),
-          ..._totalPullsWithoutRareLabels,
-          SizedBox(height: 16.h),
-          ..._pullRateLabels,
-        ],
-      ),
-    );
-  }
-
-  Widget get _titleLabel => Text(
-        title,
-        style: TextStyle(
-          color: Colors.grey[150],
-          fontSize: 22.sp,
-          fontWeight: FontWeight.bold,
-        ),
-      );
-
-  Widget get _indicators {
-    final items = Rarity.poolExclusive.map(
-      (rarity) => Consumer(
-        builder: (_, ref, __) {
-          final indicatorSize = Size(24.w, 16.h);
-          final indicator = Consumer(
-            builder: (_, ref, __) {
-              final isTouched =
-                  ref.watch(_gachaPieChartProvider).isTouched(rarity);
-              final alpha = isTouched ? 180 : 255;
-              return Container(
-                width: indicatorSize.width,
-                height: indicatorSize.height,
-                decoration: BoxDecoration(
-                  color: rarity.color.withAlpha(alpha),
-                  borderRadius: BorderRadius.circular(2),
-                ),
-              );
-            },
-          );
-          final rarityLabel = Text(
-            rarity.fullTitle,
-            style: TextStyle(
-              color: Colors.grey[130],
-              fontSize: 14.sp,
-              fontWeight: FontWeight.bold,
-              height: indicatorSize.height / 14.sp,
-            ),
-          );
-          final divider = SizedBox(width: 12.w);
-          return MouseRegion(
-            onEnter: (_) => ref.read(_gachaPieChartProvider).touch(rarity),
-            onExit: (_) => ref.read(_gachaPieChartProvider).touch(null),
-            child: Row(
-              children: [
-                indicator,
-                SizedBox(width: 4.w),
-                rarityLabel,
-                if (rarity != Rarity.poolExclusive.last) divider,
-              ],
-            ),
-          );
-        },
-      ),
-    );
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: items.toList(),
-    );
-  }
-
-  Widget get _chart => Consumer(
-        builder: (context, ref, _) {
-          final sources = stats.statsPerRarity;
-          if (sources.isEmpty) {
-            return const AppEmptyView();
-          }
-
-          final sections = stats.statsPerRarity.map((pair) {
-            final rarity = pair.first;
-            final stats = pair.second;
-            final isTouched =
-                ref.watch(_gachaPieChartProvider).isTouched(rarity);
-            final radius = isTouched ? 130.w : 120.w;
-            final fontSize = isTouched ? 14.sp : 12.sp;
-            return PieChartSectionData(
-              value: stats.chars.length.toDouble(),
-              color: rarity.color,
-              radius: radius,
-              title: this.stats.caclPullRate(rarity),
-              titleStyle: TextStyle(color: Colors.white, fontSize: fontSize),
+  Widget build(BuildContext context, WidgetRef ref) {
+    final pools = ref.watch(_filteredPools);
+    return ListView.separated(
+      itemBuilder: (_, index) {
+        final pool = pools[index];
+        return ref.watch(gachaStatsProvider(pool)).maybeWhen(
+              data: (stats) {
+                return Card(
+                  child: DefaultTextStyle.merge(
+                    style: TextStyle(
+                      color: Colors.grey[120],
+                      fontSize: 14.sp,
+                    ),
+                    child: Row(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Expanded(child: _buildOverview(pool, stats)),
+                        Expanded(child: _buildProbability(stats)),
+                        Expanded(flex: 6, child: _buildChars(stats)),
+                      ],
+                    ),
+                  ),
+                );
+              },
+              orElse: () => const SizedBox.shrink(),
             );
-          }).toList();
-          final data = PieChartData(
-            sections: sections,
-            centerSpaceRadius: 0,
-            sectionsSpace: 0,
-            borderData: FlBorderData(show: false),
-            pieTouchData: PieTouchData(
-              touchCallback: (event, response) =>
-                  _onTouch(context, ref, sources, event, response),
-            ),
-          );
-          return PieChart(data);
-        },
-      );
+      },
+      separatorBuilder: (_, __) => SizedBox(height: 12.h),
+      itemCount: pools.length,
+    );
+  }
 
-  Widget get _pullDateRangeLabel => Text(
-        stats.dateRange ?? '',
-        style: TextStyle(color: Colors.grey[100], fontSize: 14.sp),
-      );
-
-  Widget get _totalPullsLabel => Text.rich(
-        TextSpan(
-          children: [
-            const TextSpan(text: '共抽取'),
-            TextSpan(
-              text: ' ${stats.chars.length} ',
-              style: TextStyle(color: Colors.blue.lighter),
-            ),
-            const TextSpan(text: '次'),
-          ],
-          style: TextStyle(
-            color: Colors.grey[130],
-            fontSize: 15.sp,
-            fontWeight: FontWeight.bold,
+  Widget _buildOverview(String poolName, GachaStats stats) => Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Consumer(
+            builder: (_, ref, __) =>
+                ref.watch(gachaPoolFilterProvider(poolName)).maybeWhen(
+                      data: (pool) => Row(
+                        crossAxisAlignment: CrossAxisAlignment.baseline,
+                        textBaseline: TextBaseline.alphabetic,
+                        children: [
+                          Text(
+                            pool?.name ?? poolName,
+                            style: const TextStyle(fontWeight: FontWeight.bold),
+                          ),
+                          if (pool != null)
+                            Padding(
+                              padding: EdgeInsets.only(left: 2.w),
+                              child: Text(
+                                '※${pool.ruleType.label}',
+                                style: TextStyle(
+                                  color: Colors.purple.lightest,
+                                  fontSize: 12.sp,
+                                ),
+                              ),
+                            ),
+                        ],
+                      ),
+                      orElse: () => Text(
+                        poolName,
+                        style: const TextStyle(fontWeight: FontWeight.bold),
+                      ),
+                    ),
           ),
-        ),
-      );
-
-  List<Widget> get _totalPullsWithoutRareLabels => Rarity.rares
-      .map(
-        (rarity) => Text.rich(
-          TextSpan(
+          SizedBox(height: 2.h),
+          Row(
+            crossAxisAlignment: CrossAxisAlignment.baseline,
+            textBaseline: TextBaseline.alphabetic,
             children: [
-              const TextSpan(text: '已累计'),
-              TextSpan(
-                text: ' ${stats.sinceLastPull(rarity)} ',
-                style: TextStyle(color: Colors.blue.lighter),
-              ),
-              const TextSpan(text: '抽未出'),
-              TextSpan(
-                text: ' ${rarity.title} ',
-                style: TextStyle(color: rarity.color),
+              const Text('总计'),
+              SizedBox(width: 4.sp),
+              Text(
+                '${stats.chars.length}',
+                style: TextStyle(
+                  color: Colors.yellow.darkest,
+                  fontSize: 20.sp,
+                ),
               ),
             ],
-            style: TextStyle(
-              color: Colors.grey[130],
-              fontSize: 15.sp,
-              fontWeight: FontWeight.bold,
-            ),
           ),
-        ),
-      )
-      .toList();
-
-  List<Widget> get _pullRateLabels => Rarity.rares.map(
-        (rarity) {
-          final total = stats.filter(rarity).length;
-          if (total == 0) {
-            return const SizedBox.shrink();
-          }
-          final details = Row(
-            children: [
-              Expanded(flex: 2, child: Text('$total')),
-              Expanded(
-                flex: 4,
-                child: Text('平均出货次数：${stats.caclAvgPulls(rarity)}'),
-              ),
-            ],
-          );
-          return DefaultTextStyle.merge(
-            style: TextStyle(
-              color: rarity.color,
-              fontSize: 15.sp,
-              fontWeight: FontWeight.bold,
-            ),
-            child: SizedBox(
-              width: 320.w,
-              child: Row(
+          SizedBox(height: 4.h),
+          ...Rarity.rares.map(
+            (rarity) => Text.rich(
+              TextSpan(
                 children: [
-                  Text('${rarity.title}：'),
-                  SizedBox(width: 12.w),
-                  Expanded(child: details),
+                  TextSpan(
+                    text: rarity.label,
+                    style: TextStyle(color: rarity.color),
+                  ),
+                  const TextSpan(text: ' 已累计'),
+                  TextSpan(
+                    text: ' ${stats.sinceLastPull(rarity)} ',
+                    style: TextStyle(color: Colors.green.lighter),
+                  ),
+                  const TextSpan(text: '抽未出'),
                 ],
               ),
             ),
-          );
-        },
-      ).toList();
-
-  Future<void> _onTouch(
-    BuildContext context,
-    WidgetRef ref,
-    List<CharsPerRarity> sources,
-    FlTouchEvent event,
-    PieTouchResponse? response,
-  ) async {
-    //* Hover
-    final isInterestedForInteractions = event.isInterestedForInteractions;
-    final touchedSection = response?.touchedSection;
-    if (!isInterestedForInteractions || touchedSection == null) {
-      ref.read(_gachaPieChartProvider).touch(null);
-      return;
-    }
-    final index = touchedSection.touchedSectionIndex;
-    final rarity = sources.elementAtOrNull(index)?.first;
-    ref.read(_gachaPieChartProvider).touch(rarity);
-
-    //* Tap
-    if (event is FlTapUpEvent && rarity != null) {
-      _showCharsTable(context, rarity);
-    }
-  }
-
-  Future<void> _showCharsTable(BuildContext context, Rarity rarity) async {
-    final columns = [
-      DataColumn2(label: const SizedBox.shrink(), fixedWidth: 80.w),
-      const DataColumn2(label: Text('干员'), size: ColumnSize.L),
-      const DataColumn2(label: Text('抽数'), size: ColumnSize.S),
-      const DataColumn2(label: Text('寻访'), size: ColumnSize.L),
-      const DataColumn2(label: Text('获取时间'), size: ColumnSize.L),
-    ];
-
-    material.DataCell buildTextCell(String text, {TextStyle? style}) =>
-        material.DataCell(
-          DefaultTextStyle.merge(
-            style: style ??
-                TextStyle(
-                  color: Colors.grey[120],
-                  fontSize: 15.sp,
-                ),
-            child: Text(text),
           ),
-        );
+          SizedBox(height: 6.h),
+          Text(stats.startDateTime ?? '-'),
+          Text(stats.endDateTime ?? '-'),
+        ],
+      );
 
-    final rows = stats.filterWithPulls(rarity).reversed.mapIndexed(
-      (index, pair) {
-        final char = pair.first;
-        final pulls = pair.second;
-        return DataRow2.byIndex(
-          index: index,
-          cells: [
-            buildTextCell(
-              '${index + 1}',
-              style: TextStyle(color: Colors.grey[80], fontSize: 14.sp),
-            ),
-            material.DataCell(
-              AppBadge(
-                showBadge: char.isNew,
-                position: BadgePosition.topEnd(top: 4.h, end: -34.w),
-                badgeContent: Text(
-                  'NEW',
-                  style: DefaultTextStyle.of(context).style.copyWith(
-                        color: Colors.white,
-                        fontSize: 9.sp,
-                      ),
+  Widget _buildProbability(GachaStats stats) => Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Text('概率'),
+          SizedBox(height: 6.h),
+          ...Rarity.rares.map(
+            (rarity) {
+              final total = stats.filter(rarity).length;
+              return DefaultTextStyle.merge(
+                style: TextStyle(color: rarity.color),
+                child: Row(
+                  children: [
+                    Text('${rarity.label}：'),
+                    SizedBox(
+                      width: 48.w,
+                      child: Text('$total'),
+                    ),
+                    Text(
+                      '[${stats.caclPullRate(rarity)}]',
+                    ),
+                  ],
                 ),
-                child: Text.rich(
+              );
+            },
+          ),
+          SizedBox(height: 6.h),
+          ...Rarity.rares.map(
+            (rarity) => Text.rich(
+              TextSpan(
+                children: [
                   TextSpan(
-                    text: char.name,
-                    recognizer: TapGestureRecognizer()
-                      ..onTap = () => Routes.webview.push(
-                            context,
-                            extra: WebviewParams(
-                              initialUrl: '$prts/w/${char.encodeName}',
-                            ),
-                          ),
+                    text: rarity.label,
+                    style: TextStyle(color: rarity.color),
                   ),
-                  style: DefaultTextStyle.of(context).style.copyWith(
-                        color: rarity.color,
-                        fontSize: 14.sp,
-                        fontWeight: FontWeight.bold,
-                      ),
-                ),
+                  const TextSpan(
+                    text: ' 平均出货次数 ',
+                  ),
+                  TextSpan(
+                    text: stats.caclAvgPulls(rarity),
+                    style: TextStyle(color: Colors.green.lighter),
+                  ),
+                ],
               ),
             ),
-            buildTextCell('$pulls'),
-            buildTextCell(char.pool),
-            buildTextCell(char.ts.dateTime.yMMMdHmsString),
-          ],
-        );
-      },
-    ).toList();
-    final dataTable = DataTable2(
-      columns: columns,
-      headingTextStyle: DefaultTextStyle.of(context).style.copyWith(
-            fontSize: 14.sp,
-            fontWeight: FontWeight.bold,
           ),
-      rows: rows,
-    );
+          SizedBox(height: 6.h),
+          Row(
+            children: [
+              Text.rich(
+                TextSpan(
+                  children: [
+                    const TextSpan(text: '最欧 '),
+                    TextSpan(
+                      text: stats.caclMinPulls(Rarity.six),
+                      style: TextStyle(color: Colors.yellow.darkest),
+                    ),
+                  ],
+                ),
+              ),
+              SizedBox(width: 12.w),
+              Text.rich(
+                TextSpan(
+                  children: [
+                    const TextSpan(text: '最非 '),
+                    TextSpan(
+                      text: stats.caclMaxPulls(Rarity.six),
+                      style: TextStyle(color: Colors.blue.darkest),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ],
+      );
 
-    await showCustomModalBottomSheet<void>(
-      context: context,
-      builder: (_) => Container(
-        color: Colors.white,
-        height: 540.h,
-        padding: EdgeInsets.only(left: 12.w, top: 16.h, right: 12.w),
-        child: dataTable,
-      ),
-      containerWidget: (_, __, child) {
-        return SafeArea(
-          child: Padding(
-            padding: EdgeInsets.symmetric(
-              vertical: 24.h,
-              horizontal: 480.w,
-            ),
-            child: material.Material(
-              clipBehavior: Clip.antiAlias,
-              borderRadius: BorderRadius.circular(8),
-              child: child,
-            ),
+  Widget _buildChars(GachaStats stats) {
+    final charWithPulls = stats.filterCharWithPulls(Rarity.six);
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text('${Rarity.six.label}记录'),
+        SizedBox(height: 6.h),
+        if (charWithPulls.isEmpty) const Text('无'),
+        if (charWithPulls.isNotEmpty)
+          Wrap(
+            spacing: 6.sp,
+            children: charWithPulls
+                .map(
+                  (charWithPull) => Text.rich(
+                    TextSpan(
+                      children: [
+                        TextSpan(text: charWithPull.first.name),
+                        TextSpan(text: '[${charWithPull.second}]'),
+                      ],
+                      style: TextStyle(
+                        color: Rarity.exploreYourFate(charWithPull.second),
+                      ),
+                    ),
+                  ),
+                )
+                .toList(),
           ),
-        );
-      },
+      ],
     );
   }
 }
