@@ -63,7 +63,7 @@ class SettingsAboutSection extends StatelessWidget {
                 children: [
                   const TextSpan(
                     text:
-                        '$appName 是一项非盈利的个人开源项目，专注于为各位博士提供方便快捷安全的游戏数据统计和查询服务。\n'
+                        '${Constants.appName} 是一项非盈利的个人开源项目，专注于为各位博士提供方便快捷安全的游戏数据统计和查询服务。\n'
                         '由于尚处于开发阶段，在您的使用过程中不可避免地会出现一些问题，请您及时向我们反馈以便给您带来更好的使用体验。\n',
                   ),
                   const TextSpan(text: '欢迎前往我们的'),
@@ -71,7 +71,7 @@ class SettingsAboutSection extends StatelessWidget {
                     text: ' GitHub ',
                     style: TextStyle(color: Colors.blue.normal),
                     recognizer: TapGestureRecognizer()
-                      ..onTap = () => launchUrlString(projectUrl),
+                      ..onTap = () => launchUrlString(Constants.projectUrl),
                   ),
                   const TextSpan(
                     text: '为我们点亮 Star、提交 Issues 以及 Pull Requests。',
@@ -109,7 +109,7 @@ class SettingsAboutSection extends StatelessWidget {
                     text: ' BSD 3-Clause "New" or "Revised" License ',
                     style: TextStyle(color: Colors.blue.normal),
                     recognizer: TapGestureRecognizer()
-                      ..onTap = () => launchUrlString(licenseUrl),
+                      ..onTap = () => launchUrlString(Constants.licenseUrl),
                   ),
                   const TextSpan(text: '开源许可协议。'),
                 ],
@@ -175,7 +175,8 @@ class SettingsAboutSection extends StatelessWidget {
                         text: ' Kengxxiao/ArknightsGameData ',
                         style: TextStyle(color: Colors.blue.normal),
                         recognizer: TapGestureRecognizer()
-                          ..onTap = () => launchUrlString(gameDataResourceUrl),
+                          ..onTap =
+                              () => launchUrlString(Constants.gameDataRepoUrl),
                       ),
                     ],
                   ),
