@@ -3,8 +3,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:time/time.dart';
 import 'package:window_manager/window_manager.dart';
 
-import '../../application/game_data/game_data_provider.dart';
-import '../../application/game_data/states/gacha_pool_state.dart';
+import '../../application/game_data_raw/gacha_pool_provider.dart';
+import '../../application/game_data_raw/states/gacha_pool_state.dart';
 import '../../application/splash/splash_provider.dart';
 import '../core/common/widgets/app_dialog.dart';
 import '../core/common/widgets/app_flush_bar.dart';
@@ -67,7 +67,7 @@ class _SplashPageState extends ConsumerState<SplashPage> with WindowListener {
             failure.maybeWhen(
               unexpectedError: (_) async {
                 await 3.seconds.delay;
-                ref.read(gachaPoolProvider.notifier).retry();
+                ref.invalidate(gachaPoolProvider);
               },
               orElse: () {},
             );
